Big Code Cleanup
Giant code cleanup/removal of things that are no longer used.

I basically ran Android lint and followed a bunch of its suggestions.
One of the side effects of this is changing a lot of pixels to
display-independent pixels. This means that sizing is a lot better on
the Triumph, with it's larger, higher-resoultion screen.